+++ tester/remote_provisioning_tester.c
@@ -33,7 +33,7 @@ static void remote_provisioning_http(void) {
 	BC_ASSERT_TRUE(wait_for(marie->lc,NULL,&marie->stat.number_of_LinphoneConfiguringSuccessful,1));
 	BC_ASSERT_TRUE(wait_for(marie->lc,NULL,&marie->stat.number_of_LinphoneRegistrationOk,1));
 	/*make sure proxy config is not added in double, one time at core init, next time at configuring successfull*/
-	BC_ASSERT_EQUAL(bctbx_list_size(linphone_core_get_proxy_config_list(marie->lc)), 1, int,"%i");
+	BC_ASSERT_EQUAL(bctbx_list_size(linphone_core_get_proxy_config_list(marie->lc)), 1, size_t, "%zu");
 	linphone_core_manager_destroy(marie);
 }
